fix building wasm filer for mac
